+++ server/routes/auth.js
@@ -1,4 +1,5 @@
 import express from 'express';
+import multer from 'multer';
 
 import * as verifyController from '../controllers/verify';
 import * as authController from '../controllers/auth';
@@ -6,10 +7,11 @@ import * as userController from '../controllers/users';
 
 const router = express.Router();
 
+
 router.post('/sendCode', verifyController.sendCode);
 router.post('/checkCode', verifyController.checkCode, userController.createUser);
 
-router.post('/login', authController.login);
+router.post('/login', multer().array(), authController.login);
 router.post('/check', authController.checkAuth);
 
 export default router;