+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 
+# 3/12
+- Change Login API with contentType: 'multipart/form-data'
+- 
+
 # 3/11
 - Add Typescript 
 - ADD Typescript test API.